fix: main메서드와 controller 분리
